Others/release_pipelines/requirements.txt
python-gitlab>=3.0.0
python-dotenv>=1.0.0